update function & check rule severity
